+++ tamguo/src/main/java/com/tamguo/service/impl/QuestionService.java
@@ -13,6 +13,8 @@ import com.tamguo.dao.QuestionMapper;
 import com.tamguo.model.PaperEntity;
 import com.tamguo.model.QuestionEntity;
 import com.tamguo.service.IQuestionService;
+import com.tamguo.util.Result;
+import com.tamguo.util.ShiroUtils;
 
 @Service
 public class QuestionService extends ServiceImpl<QuestionMapper, QuestionEntity> implements IQuestionService{
@@ -45,10 +47,14 @@ public class QuestionService extends ServiceImpl<QuestionMapper, QuestionEntity>
 
 	@Transactional(readOnly=false)
 	@Override
-	public void addQuestion(QuestionEntity question) {
+	public Result addQuestion(QuestionEntity question) {
 		PaperEntity paper = paperMapper.selectById(question.getPaperId().toString());
+		if(!ShiroUtils.getUserId().equals(paper.getCreaterId())) {
+			return Result.result(501, null, "改试卷不属于您！");
+		}
 		question.setCourseId(paper.getCourseId());
 		questionMapper.insert(question);
+		return Result.result(0, null, "添加成功！");
 	}
 
 	@Transactional(readOnly=true)
@@ -63,14 +69,25 @@ public class QuestionService extends ServiceImpl<QuestionMapper, QuestionEntity>
 
 	@Transactional(readOnly=false)
 	@Override
-	public void updateQuestion(QuestionEntity question) {
+	public Result updateQuestion(QuestionEntity question) {
+		PaperEntity paper = paperMapper.selectById(question.getPaperId().toString());
+		if(!ShiroUtils.getUserId().equals(paper.getCreaterId())) {
+			return Result.result(501, null, "改试卷不属于您！");
+		}
 		questionMapper.updateById(question);
+		return Result.result(0, null, "修改成功！");
 	}
 
 	@Transactional(readOnly=false)
 	@Override
-	public void delete(String uid) {
+	public Result delete(String uid) {
+		QuestionEntity question = questionMapper.selectById(uid);
+		PaperEntity paper = paperMapper.selectById(question.getPaperId().toString());
+		if(!ShiroUtils.getUserId().equals(paper.getCreaterId())) {
+			return Result.result(501, null, "改试卷不属于您！");
+		}
 		questionMapper.deleteById(uid);
+		return Result.result(0, null, "删除成功！");
 	}
 
 	@Transactional(readOnly=true)
